initial fixes to separate files out of old diptych.cc
